Added pvp_ladder.txt
This is a script by Stolao from rAthena. The original versions by him doesn't work well as he was not able to actually test things.
As I was trying to use it, I found the bugs and decided to fix them for release here in Hercules as I've noticed we do not have a ladder here.

@Stolao, if you see this I am sorry for fixing your script without permission and for releasing my unofficial version of it here. Let me know if you need me to take this down and I will be happy to do so.


Features:
So basically its a simple PvP Ladder that is using arrays & player permanent variable instead of custom SQL table to store the ranking.

    GMs/Admins can reset the ranking
    Players can view the ranking
    Configurable option to give #Pvppoints or any variable for use on a shop/npc which is not included here.
    Configurable anti repeat kill abuse, disable/enable on WoE, Announce, Max Players to Rank
    @PvpLadder to easily view ranks


So yeah, All credits to Stolao, I mainly just made it work by fixing the bugs and debugging. I apologize if its not as clean as it should be, I was kinda in rush when i fixed it. I'll probably clean it up when I get some time. 

What's New in Version 2.0 (See full changelog)

    1.00 Initial Release
    1.01 Added Location with veiwing of ladder
    1.02 Added GM Menu
    1.03 Added atcommand
    1.04 Shortend menu script
    1.05 Added PvpPoints into confug and script
    1.06 Added Abuse Protection
    1.07 Made Pvp Points Configurable
    1.08 Fixed typo in Reset Option
    1.09 Made Amount of pvp points earned configurable
    1.0A Ajusted Spacing for easier reading
    1.0B Added Warp to savepoint on abuse into config
    2.0 (unofficial)Fixed bugs on ranking modification and bitwise permissions